fix(cj4): only use lbs paxWeight for zero fuel weight
paxWeight was used both for calculating the zero fuel weight, but also
for displaying the weight per passenger.  While the weight per passenger
would display correctly, it would be calculated incorrectly with a lower
weight due to all calculations being done in lbs.

This commit postpones conversion of paxWeight until displaying the
value, which seems to be the norm. It then makes sure it uses
WT_ConvertUnit when reading and comparing the paxWeight too.

Fixes #1102. Hopefully.